Fix defect when scheduled item only runs once per week
I have not been able to test this yet, but I realized I made a mistake in figureing the initial delay for scheduled items which only run once per week.  This attempts to solve that issue by adding in how many days we need to wait till that first record time occurs.